corrected the get time
svn path=/trunk/iCub/; revision=8245

+++ src/modules/selectiveAttentionEngine/include/iCub/selectiveAttentionProcessor.h
@@ -275,7 +275,7 @@ class selectiveAttentionProcessor:public yarp::os::RateThread {
         /**
         * function that returns the value of the saccadicInterval
         */
-        double getSaccadicInterval() { return saccadicInterval; };
+        int getSaccadicInterval() { return saccadicInterval; };
 
         /**
         * function that returns the value of the parameter k1
